Move `Toggle` to `design` package
